+++ src/Ai/Base/Actions/TravelAction.cpp
@@ -37,7 +37,7 @@ bool TravelAction::Execute(Event /*event*/)
         if (!newTarget->IsAlive())
             continue;
 
-        if (newTarget->GetEntry() == target->getDestination()->getEntry())
+        if (newTarget->GetEntry() == uint32(target->getDestination()->getEntry()))
             continue;
 
         if (newTarget->IsInCombat())
