+++ src/Ai/Raid/Naxx/Action/NaxxActions_Razuvious.cpp
@@ -47,7 +47,7 @@ bool RazuviousUseObedienceCrystalAction::Execute(Event /*event*/)
         {
             // taunt
             bool tauntUseful = true;
-            if (forceObedience->GetDuration() <= (duration_time - 5000))
+            if (forceObedience->GetDuration() <= int32(duration_time - 5000))
             {
                 Unit* victim = target->GetVictim();
                 if (victim && victim->HasAura(SPELL_BONE_BARRIER))
@@ -57,7 +57,7 @@ bool RazuviousUseObedienceCrystalAction::Execute(Event /*event*/)
                     tauntUseful = false;
 
             }
-            if (forceObedience->GetDuration() >= (duration_time - 500))
+            if (forceObedience->GetDuration() >= int32(duration_time - 500))
                 tauntUseful = false;
 
             if (tauntUseful && !charm->HasSpellCooldown(29060))
