+++ src/Bot/Engine/Trigger/Trigger.cpp
@@ -36,7 +36,7 @@ bool Trigger::needCheck(uint32 now)
     if (checkInterval < 2)
         return true;
 
-    if (!lastCheckTime || now - lastCheckTime >= checkInterval)
+    if (!lastCheckTime || now - lastCheckTime >= uint32(checkInterval))
     {
         lastCheckTime = now;
         return true;
